update base lib

+++ examples/timer/main.cpp
@@ -8,30 +8,33 @@
 
 using namespace std::chrono_literals;
 
-cppevent::task<void> signal_coroutine(cppevent::event_listener& listener, bool& control) {
-    while (control) {
+cppevent::task<void> signal_coroutine(cppevent::event_listener& listener) {
+    for (int i = 1; i <= 5; ++i) {
         co_await cppevent::read_awaiter { listener };
-        std::cout << "Received an in-app signal" << std::endl;
+        std::cout << "Received an in-app signal: " << i << std::endl;
     }
 }
 
-cppevent::task<void> timed_coroutine(cppevent::event_loop& e_loop, bool& control) {
+cppevent::task<void> timed_coroutine(cppevent::event_loop& e_loop) {
+    bool control = true;
     auto& signal_listener = *(e_loop.get_signal_listener());
-    auto t = signal_coroutine(signal_listener, control);
+    auto t = signal_coroutine(signal_listener);
     cppevent::timer timer(3s, e_loop);
     std::cout << "Starting timed coroutine which triggers every 3 secs 5 times" << std::endl;
-    for (int i = 1; i <= 5; ++i) {
+    while (true) {
         co_await timer.wait();
-        std::cout << "Timer triggered: " << i << std::endl;
+        if (t.m_handle.done()) {
+            break;
+        }
+        std::cout << "Timer triggered" << std::endl;
         e_loop.send_signal(&signal_listener, true, false);
     }
-    control = false;
+    e_loop.stop();
 }
 
 int main() {
     cppevent::event_loop e_loop;
-    bool control = true;
-    timed_coroutine(e_loop, control);
-    e_loop.run(control);
+    timed_coroutine(e_loop);
+    e_loop.run();
     return 0;
 }
